Pannels
Note the use of the panel and panel-primary classes to style the panel. The panel contains two divs, one with panel-heading class to host the heading of the panel, and the second with the panel-body class to host the content. Note the use of <dl> tag with the dl-horizontal class to create a horizontal description list.